Add files via upload
This project is the second challenge project from Full-Stack Engineer Career Path on Codecademy learning platform. In this project, I built my own basic design system for a website. In essence, this is a website to help you build MORE websites in the future! On your site, are collected all the colors, fonts, and some of the repeating styles.